refactor(utf8-safe): split Resume.jsx into modular files under 200 lines
- Extract design tokens to src/tokens.js (37 lines)
- Extract layout styled-components to src/styled.js (89 lines)
- Extract sidebar styled-components to src/styled-sidebar.js (104 lines)
- Extract main-column styled-components to src/styled-main.js (102 lines)
- Extract sidebar sections to src/sections/sidebar.jsx (143 lines)
- Extract main sections to src/sections/main.jsx (193 lines)
- Resume.jsx reduced from 901 to 78 lines (composition only)
- Fix: remove duplicate sections (certs/languages/interests only in sidebar)
- Fix: wrap all external URLs with safeUrl()
- Fix: use stable React keys (not key={i})
- Regenerate pnpm-lock.yaml

Fixes CodeRabbit issues #2 (file size) and #3 (duplicate sections).
