+++ packages/themes/jsonresume-theme-utf8-safe/src/tokens.js
@@ -1,0 +1,37 @@
+/*
+ * UTF-8 Safe — Design Tokens
+ *
+ * SAP OpenUI5 Horizon inspired palette and spacing.
+ */
+
+export const colors = {
+  primary: '#0070f2',
+  primaryHover: '#005dc9',
+  primaryLight: '#e1effa',
+  bg: '#f5f6f7',
+  surface: '#ffffff',
+  text: '#1d2a3a',
+  textSecondary: '#475e75',
+  textSubtle: '#6a7e92',
+  border: '#d9e1e8',
+  borderLight: '#e8edf1',
+  sidebarBg: '#fafbfc',
+  headerBg: '#1d2a3a',
+  headerText: '#ffffff',
+  headerSubtle: '#a8b8c9',
+  tagBg: '#eef1f5',
+  tagText: '#475e75',
+  tagAccentBg: '#e1effa',
+  tagAccentText: '#0070f2',
+  marker: '#0070f2',
+  markerLine: '#d9e1e8',
+  link: '#0070f2',
+};
+
+export const spacing = {
+  xs: '4px',
+  sm: '8px',
+  md: '16px',
+  lg: '24px',
+  xl: '32px',
+};
